# 0.3.2.1
    - add the edition of folder name functionality
        + added edit button
        - to create edit form
    + added the delete folder functionality
    + added the window confirming deletion for folder and bookmark

+++ src/main/resources/static/js/div-id-content.js
@@ -429,7 +429,9 @@ $(document).ready(function () {
                     let bookmarkDelBtn = $('#delete-bookmark-btn'+e.id);
                     bookmarkDelBtn.on("click", function (){
                        //alert("test delete bookmark id="+e.id+" button!"); // ok
-                        deleteBookmark(e.id, folderId);
+                        if (confirm("Confirm deletion...")) {
+                            deleteBookmark(e.id, folderId);
+                        }
 
                     });
                 });
